Committed by anchal

+++ src/main/java/cc/altius/FASP/model/rowMapper/NewSupplyPlanRegionResultSetExtractor.java
@@ -34,41 +34,41 @@ public class NewSupplyPlanRegionResultSetExtractor implements ResultSetExtractor
             RegionData rd = new RegionData();
             rd.setRegionId(rs.getInt("REGION_ID"));
             if (rs.wasNull()) {
-                nsp.addPlannedShipmentsTotalData(rs.getInt("MANUAL_PLANNED_SHIPMENT"));
-                nsp.addSubmittedShipmentsTotalData(rs.getInt("MANUAL_SUBMITTED_SHIPMENT"));
-                nsp.addApprovedShipmentsTotalData(rs.getInt("MANUAL_APPROVED_SHIPMENT"));
-                nsp.addShippedShipmentsTotalData(rs.getInt("MANUAL_SHIPPED_SHIPMENT"));
-                nsp.addReceivedShipmentsTotalData(rs.getInt("MANUAL_RECEIVED_SHIPMENT"));
-                nsp.addOnholdShipmentsTotalData(rs.getInt("MANUAL_ONHOLD_SHIPMENT"));
-                nsp.addPlannedErpShipmentsTotalData(rs.getInt("ERP_PLANNED_SHIPMENT"));
-                nsp.addSubmittedErpShipmentsTotalData(rs.getInt("ERP_SUBMITTED_SHIPMENT"));
-                nsp.addApprovedErpShipmentsTotalData(rs.getInt("ERP_APPROVED_SHIPMENT"));
-                nsp.addShippedErpShipmentsTotalData(rs.getInt("ERP_SHIPPED_SHIPMENT"));
-                nsp.addReceivedErpShipmentsTotalData(rs.getInt("ERP_RECEIVED_SHIPMENT"));
-                nsp.addOnholdErpShipmentsTotalData(rs.getInt("ERP_ONHOLD_SHIPMENT"));
+                nsp.addPlannedShipmentsTotalData(rs.getLong("MANUAL_PLANNED_SHIPMENT"));
+                nsp.addSubmittedShipmentsTotalData(rs.getLong("MANUAL_SUBMITTED_SHIPMENT"));
+                nsp.addApprovedShipmentsTotalData(rs.getLong("MANUAL_APPROVED_SHIPMENT"));
+                nsp.addShippedShipmentsTotalData(rs.getLong("MANUAL_SHIPPED_SHIPMENT"));
+                nsp.addReceivedShipmentsTotalData(rs.getLong("MANUAL_RECEIVED_SHIPMENT"));
+                nsp.addOnholdShipmentsTotalData(rs.getLong("MANUAL_ONHOLD_SHIPMENT"));
+                nsp.addPlannedErpShipmentsTotalData(rs.getLong("ERP_PLANNED_SHIPMENT"));
+                nsp.addSubmittedErpShipmentsTotalData(rs.getLong("ERP_SUBMITTED_SHIPMENT"));
+                nsp.addApprovedErpShipmentsTotalData(rs.getLong("ERP_APPROVED_SHIPMENT"));
+                nsp.addShippedErpShipmentsTotalData(rs.getLong("ERP_SHIPPED_SHIPMENT"));
+                nsp.addReceivedErpShipmentsTotalData(rs.getLong("ERP_RECEIVED_SHIPMENT"));
+                nsp.addOnholdErpShipmentsTotalData(rs.getLong("ERP_ONHOLD_SHIPMENT"));
                 nsp.setRegionCountForStock(rs.getInt("REGION_STOCK_COUNT"));
                 nsp.setRegionCount(rs.getInt("REGION_COUNT"));
             } else {
                 nsp.setActualConsumptionFlag(rs.getBoolean("USE_ACTUAL_CONSUMPTION"));
                 nsp.setRegionCountForStock(rs.getInt("REGION_STOCK_COUNT"));
                 nsp.setRegionCount(rs.getInt("REGION_COUNT"));
-                rd.setForecastedConsumption(rs.getInt("FORECASTED_CONSUMPTION"));
+                rd.setForecastedConsumption(rs.getLong("FORECASTED_CONSUMPTION"));
                 if (rs.wasNull()) {
                     rd.setForecastedConsumption(null);
                 }
                 nsp.addForecastedConsumptionQty(rd.getForecastedConsumption());
-                rd.setActualConsumption(rs.getInt("ACTUAL_CONSUMPTION"));
+                rd.setActualConsumption(rs.getLong("ACTUAL_CONSUMPTION"));
                 if (rs.wasNull()) {
                     rd.setActualConsumption(null);
                 }
                 nsp.addActualConsumptionQty(rd.getActualConsumption());
                 nsp.addFinalConsumptionQty(nsp.isActualConsumptionFlag() ? rd.getActualConsumption() : rd.getForecastedConsumption());
-                rd.setStock(rs.getInt("STOCK"));
+                rd.setStock(rs.getLong("STOCK"));
                 if (rs.wasNull()) {
                     rd.setStock(null);
                 }
                 nsp.addStockQty(rd.getStock());
-                rd.setAdjustment(rs.getInt("ADJUSTMENT"));
+                rd.setAdjustment(rs.getLong("ADJUSTMENT"));
                 if (rs.wasNull()) {
                     rd.setAdjustment(null);
                 }
